refactor: move StreamParams from common to config
