+++ src/service/alerts/mod.rs
@@ -18,7 +18,7 @@ use arrow_schema::DataType;
 use chrono::{Duration, Utc};
 use config::{
     get_config, ider,
-    meta::search::SearchEventType,
+    meta::{search::SearchEventType, stream::StreamParams},
     utils::{
         base64,
         json::{Map, Value},
@@ -27,9 +27,8 @@ use config::{
 
 use super::promql;
 use crate::{
-    common::meta::{
-        alerts::{AggFunction, Condition, Operator, QueryCondition, QueryType, TriggerCondition},
-        stream::StreamParams,
+    common::meta::alerts::{
+        AggFunction, Condition, Operator, QueryCondition, QueryType, TriggerCondition,
     },
     service::search as SearchService,
 };
