+++ src/service/db/file_list/local.rs
@@ -16,7 +16,7 @@
 use std::io::{BufRead, BufReader};
 
 use config::{
-    meta::stream::{FileKey, FileMeta, StreamType},
+    meta::stream::{FileKey, FileMeta, StreamParams, StreamType},
     utils::{
         asynchronism::file::get_file_contents, file::scan_files, json,
         parquet::parse_file_key_columns,
@@ -26,7 +26,7 @@ use infra::file_list as infra_file_list;
 use once_cell::sync::Lazy;
 use tokio::sync::RwLock;
 
-use crate::common::{infra::wal, meta::stream::StreamParams};
+use crate::common::infra::wal;
 
 /// use queue to batch send broadcast to other nodes
 pub static BROADCAST_QUEUE: Lazy<RwLock<Vec<FileKey>>> =
