+++ src/service/logs/bulk.rs
@@ -24,7 +24,7 @@ use chrono::{Duration, Utc};
 use config::{
     get_config,
     meta::{
-        stream::{Routing, StreamType},
+        stream::{Routing, StreamParams, StreamType},
         usage::UsageType,
     },
     metrics,
@@ -36,7 +36,6 @@ use crate::{
     common::meta::{
         functions::{StreamTransform, VRLResultResolver},
         ingestion::{BulkResponse, BulkResponseError, BulkResponseItem, IngestionStatus},
-        stream::StreamParams,
     },
     service::{
         format_stream_name, ingestion::check_ingestion_allowed, schema::get_upto_discard_error,
