+++ src/service/logs/otlp_grpc.rs
@@ -22,7 +22,7 @@ use chrono::{Duration, Utc};
 use config::{
     get_config,
     meta::{
-        stream::{Routing, StreamType},
+        stream::{Routing, StreamParams, StreamType},
         usage::UsageType,
     },
     metrics,
@@ -38,7 +38,6 @@ use crate::{
     common::meta::{
         functions::{StreamTransform, VRLResultResolver},
         ingestion::{IngestionStatus, StreamStatus},
-        stream::StreamParams,
     },
     handler::http::request::CONTENT_TYPE_PROTO,
     service::{
