+++ src/service/logs/otlp_http.rs
@@ -22,7 +22,7 @@ use chrono::{Duration, Utc};
 use config::{
     get_config,
     meta::{
-        stream::{Routing, StreamType},
+        stream::{Routing, StreamParams, StreamType},
         usage::UsageType,
     },
     metrics,
@@ -39,7 +39,6 @@ use crate::{
         functions::{StreamTransform, VRLResultResolver},
         http::HttpResponse as MetaHttpResponse,
         ingestion::{IngestionStatus, StreamStatus},
-        stream::StreamParams,
     },
     handler::http::request::CONTENT_TYPE_JSON,
     service::{
