+++ src/service/logs/syslog.rs
@@ -23,7 +23,7 @@ use anyhow::Result;
 use chrono::{Duration, Utc};
 use config::{
     get_config,
-    meta::stream::{Routing, StreamType},
+    meta::stream::{Routing, StreamParams, StreamType},
     metrics,
     utils::{flatten, json},
 };
@@ -37,7 +37,6 @@ use crate::{
             functions::{StreamTransform, VRLResultResolver},
             http::HttpResponse as MetaHttpResponse,
             ingestion::{IngestionResponse, IngestionStatus, StreamStatus},
-            stream::StreamParams,
             syslog::SyslogRoute,
         },
     },
