+++ src/service/search/mod.rs
@@ -23,7 +23,7 @@ use config::{
     meta::{
         cluster::RoleGroup,
         search,
-        stream::{FileKey, StreamType},
+        stream::{FileKey, StreamParams, StreamType},
         usage::{RequestStats, UsageType},
     },
     metrics,
@@ -51,8 +51,7 @@ use {
 
 use super::usage::report_request_usage_stats;
 use crate::{
-    common::{infra::cluster as infra_cluster, meta::stream::StreamParams},
-    handler::grpc::request::search::Searcher,
+    common::infra::cluster as infra_cluster, handler::grpc::request::search::Searcher,
     service::format_partition_key,
 };
